+++ src/main/java/com/carbonldp/apps/roles/SesameAppRoleRepository.java
@@ -35,6 +35,7 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 
 	private String containerSlug;
 	private String agentsContainerSlug;
+
 	private static String getParentsQuery;
 
 	private final ContainerDescription.Type appRolesContainerType = ContainerDescription.Type.BASIC;
@@ -56,6 +57,16 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 			.append( RDFNodeUtil.generatePredicateStatement( "?members", "?agent", AppRoleDescription.Property.AGENT ) )
 		;
 		getByAgent_query = queryBuilder.toString();
+	}
+
+	public void setAppRoleContainerSlug( String slug ) {
+		Assert.notNull( slug );
+		this.containerSlug = slug;
+	}
+
+	public void setAgentsContainerSlug( String slug ) {
+		Assert.notNull( slug );
+		this.agentsContainerSlug = slug;
 	}
 
 	@Override
@@ -97,16 +108,6 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 	}
 
 	@Override
-	public AppRole create( AppRole appRole ) {
-		URI containerURI = getContainerURI();
-		containerRepository.createChild( containerURI, appRole );
-
-		Container agentsContainer = createAgentsContainer( appRole );
-		// TODO: Decide. Should the ACL be created here?
-		return appRole;
-	}
-
-	@Override
 	public Container createAppRolesContainer( URI rootContainerURI ) {
 		URI appRolesContainerURI = getContainerURI( rootContainerURI );
 		BasicContainer appRolesContainer = BasicContainerFactory.getInstance().create( new RDFResource( appRolesContainerURI ) );
@@ -114,17 +115,12 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 		return appRolesContainer;
 	}
 
-	public void setAppRoleContainerSlug( String slug ) {
-		Assert.notNull( slug );
-		this.containerSlug = slug;
+	@Override
+	public boolean exists( URI appRoleURI ) {
+		return sourceRepository.exists( appRoleURI );
 	}
 
-	public void setAgentsContainerSlug( String slug ) {
-		Assert.notNull( slug );
-		this.agentsContainerSlug = slug;
-	}
-
-	private URI getContainerURI() {
+	public URI getContainerURI() {
 		AppContext appContext = AppContextHolder.getContext();
 		if ( appContext.isEmpty() ) throw new IllegalStateException( "The rootContainerURI cannot be retrieved from the platform context." );
 		URI rootContainerURI = appContext.getApplication().getRootContainerURI();
@@ -136,15 +132,7 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 		return URIUtil.createChildURI( rootContainerURI, containerSlug );
 	}
 
-	private Container createAgentsContainer( AppRole appRole ) {
-		URI agentsContainerURI = getAgentsContainerURI( appRole.getURI() );
-		RDFResource resource = new RDFResource( agentsContainerURI );
-		DirectContainer container = DirectContainerFactory.getInstance().create( resource, appRole.getURI(), AppRoleDescription.Property.AGENT.getURI() );
-		sourceRepository.createAccessPoint( appRole.getURI(), container );
-		return container;
-	}
-
-	private URI getAgentsContainerURI( URI appRoleURI ) {
+	public URI getAgentsContainerURI( URI appRoleURI ) {
 		return URIUtil.createChildURI( appRoleURI, agentsContainerSlug );
 	}
 
@@ -170,6 +158,11 @@ public class SesameAppRoleRepository extends AbstractSesameLDPRepository impleme
 
 			return parents;
 		} );
+	}
+
+	@Override
+	public void delete( URI appRoleURI ) {
+		sourceRepository.delete( appRoleURI );
 	}
 
 }