+++ src/main/java/com/carbonldp/apps/roles/web/AppRolePUTHandler.java
@@ -1,12 +1,13 @@
 package com.carbonldp.apps.roles.web;
 
+import com.carbonldp.apps.AppRole;
 import com.carbonldp.apps.roles.AppRoleService;
 import com.carbonldp.exceptions.InvalidResourceException;
 import com.carbonldp.ldp.containers.AbstractPUTRequestHandler;
 import com.carbonldp.ldp.containers.AddMembersAction;
 import com.carbonldp.ldp.containers.AddMembersActionFactory;
+import com.carbonldp.ldp.containers.MembersAction;
 import com.carbonldp.models.Infraction;
-import com.carbonldp.rdf.RDFResource;
 import com.carbonldp.web.RequestHandler;
 import org.openrdf.model.URI;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,21 +15,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.util.List;
 
 /**
- * @author NestorEstrada
- * @since 0.19.0-ALPHA
+ * @author NestorVenegas
+ * @since 0.18.0-ALPHA
  */
 @RequestHandler
-public class AppRolesPUTAgentsHandler extends AbstractPUTRequestHandler<RDFResource> {
+public class AppRolePUTHandler extends AbstractPUTRequestHandler<AppRole> {
 
-	protected AppRoleService appRoleService;
+	AppRoleService appRoleService;
 
 	@Override
 	protected void executeAction( URI targetUri, AddMembersAction members ) {
 		validate( members );
-		appRoleService.addAgents( targetUri, members.getMembers() );
+		appRoleService.addChildren( targetUri, members.getMembers() );
 	}
 
-	protected void validate( AddMembersAction membersAction ) {
+	protected void validate( MembersAction membersAction ) {
 		List<Infraction> infractions = AddMembersActionFactory.getInstance().validate( membersAction );
 		if ( ! infractions.isEmpty() ) throw new InvalidResourceException( infractions );
 	}