+++ src/main/java/com/carbonldp/apps/roles/web/AppRolesController.java
@@ -4,7 +4,6 @@ import com.carbonldp.descriptions.APIPreferences;
 import com.carbonldp.ldp.web.AbstractLDPController;
 import com.carbonldp.rdf.RDFDocument;
 import com.carbonldp.web.config.InteractionModel;
-import com.carbonldp.web.exceptions.NotImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -16,21 +15,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @Controller
+@RequestMapping( value = "apps/*/roles/" )
 public class AppRolesController extends AbstractLDPController {
 
-	private AppRolesPUTAgentsHandler appRolesPUTAgentsHandler;
 	private AppRolesDELETEAgentsHandler appRolesDELETEAgentsHandler;
+	private AppRolesPOSTHandler postHandler;
 
-	@RequestMapping( method = RequestMethod.POST, value = "apps/*/roles/" )
-	public void createAppRole() {
-		// TODO: Implement it
-		throw new NotImplementedException();
-	}
-
-	@RequestMapping( method = RequestMethod.PUT, value = "apps/*/roles/*/agents/" )
-	@InteractionModel( APIPreferences.InteractionModel.CONTAINER )
-	public ResponseEntity<Object> addAgentToRole( @RequestBody RDFDocument requestDocument, HttpServletRequest request, HttpServletResponse response ) {
-		return appRolesPUTAgentsHandler.handleRequest( requestDocument, request, response );
+	@RequestMapping( method = RequestMethod.POST )
+	public ResponseEntity<Object> createAppRole( @RequestBody RDFDocument requestDocument, HttpServletRequest request, HttpServletResponse response ) {
+		return postHandler.handleRequest( requestDocument, request, response );
 	}
 
 	@RequestMapping( method = RequestMethod.DELETE, value = "apps/*/roles/*/agents/" )
@@ -40,9 +33,8 @@ public class AppRolesController extends AbstractLDPController {
 	}
 
 	@Autowired
-	public void setAppRolesPUTAgentsHandler( AppRolesPUTAgentsHandler appRolesPUTAgentsHandler ) {this.appRolesPUTAgentsHandler = appRolesPUTAgentsHandler;}
-
-	@Autowired
 	public void setAppRolesDELETEAgentsHandler( AppRolesDELETEAgentsHandler appRolesDELETEAgentsHandler ) {this.appRolesDELETEAgentsHandler = appRolesDELETEAgentsHandler;}
 
+	@Autowired
+	public void setAppRolesPOSTHandler( AppRolesPOSTHandler appRolesPostHandler ) {this.postHandler = appRolesPostHandler;}
 }