+++ src/main/java/com/carbonldp/ldp/sources/SesameRDFSourceRepository.java
@@ -9,8 +9,10 @@ import com.carbonldp.rdf.RDFResourceRepository;
 import com.carbonldp.repository.DocumentGraphQueryResultHandler;
 import com.carbonldp.repository.GraphQueryResultHandler;
 import com.carbonldp.utils.RDFNodeUtil;
+import com.carbonldp.utils.SPARQLUtil;
 import com.carbonldp.utils.ValueUtil;
 import com.carbonldp.web.exceptions.NotImplementedException;
+import org.apache.commons.lang3.text.StrSubstitutor;
 import org.joda.time.DateTime;
 import org.openrdf.model.Resource;
 import org.openrdf.model.URI;
@@ -24,11 +26,12 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.carbonldp.Consts.NEW_LINE;
-import static com.carbonldp.Consts.TAB;
+import static com.carbonldp.Consts.*;
 
 @Transactional
 public class SesameRDFSourceRepository extends AbstractSesameLDPRepository implements RDFSourceRepository {
+
+	private static String isQuery;
 
 	public SesameRDFSourceRepository( SesameConnectionFactory connectionFactory, RDFResourceRepository resourceRepository, RDFDocumentRepository documentRepository ) {
 		super( connectionFactory, resourceRepository, documentRepository );
@@ -200,21 +203,20 @@ public class SesameRDFSourceRepository extends AbstractSesameLDPRepository imple
 		}
 	}
 
-	private static final String isQuery;
-
 	static {
-		isQuery = "ASK  { ?resource <" + RDF.TYPE + ">  ?type }";
+		isQuery = "ASK { ?resource " + LESS_THAN + RDF.TYPE + MORE_THAN + " ?rdfType." + "${values} }";
 	}
 
 	@Override
 	public boolean is( URI resourceURI, RDFNodeEnum type ) {
-		for ( URI typeURI : type.getURIs() ) {
-			Map<String, Value> bindings = new LinkedHashMap<>();
-			bindings.put( "resource", resourceURI );
-			bindings.put( "type", typeURI );
-			if ( sparqlTemplate.executeBooleanQuery( isQuery, bindings ) ) return true;
-		}
-		return false;
+		Map<String, Value> bindings = new LinkedHashMap<>();
+		bindings.put( "resource", resourceURI );
+
+		Map<String, String> values = new HashMap<>();
+		values.put( "values", SPARQLUtil.assignVar( "rdfType", type ) );
+		StrSubstitutor sub = new StrSubstitutor( values, "${", "}" );
+
+		return sparqlTemplate.executeBooleanQuery( sub.replace( isQuery ), bindings );
 	}
 
 	@Override