+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# 0.16.0
+# 0.21.0 (2015/12/17)
+* Agents can now be removed from an AppRole
+# 0.20.0 (2015/12/17)
+* AppRoles can now be deleted
+# 0.19.0 (2015/12/11)
+* Existing agents can now be added as members of an appRole
+# 0.18.0 (2015/12/10)
+* AppRoles can now be created and related among themselves
+# 0.17.0 (2015/12/10)
+* ACLs can be retrieved
+# 0.16.0 (2015/12/08)
 * ACLs can now be modified
 # 0.15.4 (2015/11/20)
 * FIX CORS requests not returning header Access-Control-Expose-Headers